Initial commit
-> create_micro_pngs.py

Working script which generates the microstructural image used in COMSOL using
the (x, y, R) of the list of the circles.

While COMSOL could generate these, this file allows for the flexibility for:
- generating arbitrarily high or low resolution PNGs
- no antialiasing
- can modify the pore- and particle-phase colors

-> create_col_map.py

Working on refactoring the code from Jupyter Notebook. The goal is to be reduce
the compute time when generating these images.