Initial commit for 'preprocess_ml_data.py' script
This script parses the extracted microstructure, activation, and colormap
images to create a TensorFlow pipeline.

This approach was selected compared to loading data in-memory as it bypasses
the image size limitation since images are loaded in batches.

COMPLETED:
Pipeline seems to work for loading the input, activation, and target images.

TODO:
- Further pipeline optimizations during runtime
- Load metadata into the pipeline. This is problematic due to how data is
  stored. For some reason, Tensorflow does not want to index dictionaries.