MCP-2 Refactor common functions
To prepare the story, namely extracting images from a microstructural image
and predicting within each circle, several functions/methods were factored
from the create_ml_dataset.py into the utils module for code reuse for this
story.

The script was ran and confirmed that the colors were being centered in
the circle of interest.

The measure_porosity method was not tested, though minimal changes were made
so breaking changes are not expected.